fix GCC -Wall/-Wextra/-Wpedantic compiler warnings
fixed bugs:
- bullsquid not reacting to being attacked while hunting a headcrab
- attempt to play a null sentence
- wrong reload anim for body guard when using mp5
- a mystery bug with the base grunt AI (probably something rare)
- maybe some random crashes caused by violating strict aliasing

+++ dlls/monster/CGonome.cpp
@@ -305,7 +305,7 @@ void CGonome::Precache()
 	PRECACHE_SOUND_ARRAY(pDieSounds);
 
 	// not affected by mp_soundvariety (but should be)
-	for (int i = 0; i < ARRAYSIZE(pEventSounds); i++) \
+	for (int i = 0; i < (int)ARRAYSIZE(pEventSounds); i++) \
 		PRECACHE_SOUND((char*)pEventSounds[i]);
 
 	PRECACHE_SOUND(GRAB_BLOOD_SOUND);
